+++ packages/client/src/hooks/useNotification.ts
@@ -1,10 +1,9 @@
 import { useState, useCallback } from 'react'
-import { message } from 'antd'
 
 export type PermissionState = NotificationPermission | 'unsupported'
 
 const UNSUPPORTED = 'unsupported'
-const STORAGE_KEY = 'notifications_enabled'
+const NOTIFICATION_ENABLED_KEY = 'notifications_enabled'
 
 function isSupported(): boolean {
   return typeof window !== 'undefined' && 'Notification' in window
@@ -16,7 +15,7 @@ function getPermission(): PermissionState {
 
 function loadEnabled(): boolean {
   try {
-    return localStorage.getItem(STORAGE_KEY) !== 'false'
+    return localStorage.getItem(NOTIFICATION_ENABLED_KEY) !== 'false'
   } catch {
     return true
   }
@@ -26,22 +25,18 @@ export function useNotification() {
   const [permission, setPermission] = useState<PermissionState>(getPermission)
   const [enabled, setEnabled] = useState(loadEnabled)
 
-  const requestPermission = useCallback(async (): Promise<PermissionState> => {
+  const requestPermission = async (): Promise<PermissionState> => {
     if (!isSupported()) return UNSUPPORTED
 
     const result = await Notification.requestPermission()
     setPermission(result)
     return result
-  }, [])
+  }
 
-  const toggleEnabled = useCallback((value: boolean) => {
+  const toggleEnabled = (value: boolean) => {
     setEnabled(value)
-    try {
-      localStorage.setItem(STORAGE_KEY, String(value))
-    } catch {
-      message.warning('Не удалось сохранить настройку уведомлений')
-    }
-  }, [])
+    localStorage.setItem(NOTIFICATION_ENABLED_KEY, String(value))
+  }
 
   const show = useCallback(
     (title: string, options?: NotificationOptions) => {
